Added Day 790 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
Story of our second hit on Obltelecom, after the first one on December 1, 23:
Provider: We're under a DDoS attack, but chill customers, give us two hours, we got this.
Customers: It's been five hours. Where's the internet?
Provider: Oops, we only managed to fend off 60 GB of traffic when they hit us with 120.
Customers: No internet for three days now...
Now Obltelecom surely got the experience, the experience not to promise a quick recovery from IT ARMY.
Big ups to all who joined in!
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
Історія нашої другої атаки на провайдера Облтелеком, після першої 1го грудня 2023:
Провайдер: Ми під ДДОС атакою, але шановні клієнти не хвилюйтесь, дайте дві години, в нас є позитивний досвід
Клієнти: Вже пройшло п'ять годин. Де інтернет?
Провайдер : Ой, ми тільки відбили 60 Гб трафіку, як на нас пустили 120
Клієнти: Інтернету нема вже три дні...
Тепер в Облтелекома точно буде досвід, досвід не обіцяти своїм клієнтам швидко відбитись від IT ARMY.
Красунчики всі хто приєднались!