Added Day 803 - Targets and Days (TL) Eng+Ukr
Simply run this for as long as you can to help flood Russia in the most legal, yet effective way possible 
New targets imported from db1000n:
To keep up with targets we're going to use the db1000n targets direct from their GitHub repository. 
These updates will be daily, so if the db1000n changes, it will probably take a few hours longer for the change to make it here.
Message posted by the IT Army of Ukraine:
🔥 Week #17 of 2024: IT ARMY Operations Leaders 🏆
Our intelligence has strengthened over the past week, meaning less and less of the enemy's digital infrastructure will be operational. Congratulations to those who have generated the most power:
🛠 Db1000n: Xiotak 📊 25.0 TB
🛠 Distress: frost1 📊 8,497.9 TB
🛠 Mhddos: UkrByte 📊 1,469.1 TB
🛠 X100: UkrByte 📊 1,871.4 TB
Different tools generate different amounts of traffic for achieving the desired effect, so direct comparisons are not advisable. 🛡
/ *** / 
Просто запускайте це стільки, скільки зможете, щоб допомогти наповнити Росію найбільш законним, але ефективним способом 
Нові цілі, імпортовані з db1000n:
🔥 Тиждень #17 2024: Лідери операцій IT ARMY 🏆
Наша розвідка стала сильніше минулого тижня, а значить все менше цифрової інфраструктури буде працювати у ворога. Вітаємо тих хто створив найбільше потужностей:
🛠 Db1000n: Xiotak 📊 25.0 TB
🛠 Distress: frost1 📊 8,497.9 TB
🛠 Mhddos: UkrByte 📊 1,469.1 TB
🛠 X100: UkrByte 📊 1,871.4 TB
Різні інструменти генерують різну кількість трафіку для досягнення потрібного ефекту, тож не варто їх порівнювати. 🛡